+++ Source/SwiftLintFramework/Rules/ConditionalReturnsOnNewline.swift
@@ -23,22 +23,40 @@ public struct ConditionalReturnsOnNewline: ConfigurationProviderRule, Rule, OptI
             "guard true else {\n return true\n}",
             "guard true,\n let x = true else {\n return true\n}",
             "if true else {\n return true\n}",
-            "if true,\n let x = true else {\n return true\n}"
+            "if true,\n let x = true else {\n return true\n}",
+            "if textField.returnKeyType == .Next {",
+            "if true { // return }",
+            "/*if true { */ return }"
         ],
         triggeringExamples: [
             "guard true else { return }",
             "if true { return }",
             "if true { break } else { return }",
+            "if true { break } else {       return }",
+            "if true { return \"YES\" } else { return \"NO\" }"
         ]
     )
 
     public func validateFile(_ file: File) -> [StyleViolation] {
-        let pattern = "(guard|if)[^\n]*return[^\n]\n*"
-        let excludingKinds = SyntaxKind.commentAndStringKinds()
-        return file.matchPattern(pattern, excludingSyntaxKinds: excludingKinds).map {
+        let pattern = "(guard|if)[^\n]*return"
+        return file.rangesAndTokensMatching(pattern).filter { range, tokens in
+            guard let firstToken = tokens.first, let lastToken = tokens.last,
+                SyntaxKind(rawValue: firstToken.type) == .Keyword &&
+                    SyntaxKind(rawValue: lastToken.type) == .Keyword else {
+                        return false
+            }
+
+            return ["if", "guard"].contains(contentForToken(token: firstToken, file: file)) &&
+                contentForToken(token: lastToken, file: file) == "return"
+        }.map {
             StyleViolation(ruleDescription: type(of: self).description,
                 severity: self.configuration.severity,
-                location: Location(file: file, byteOffset: $0.location))
+                location: Location(file: file, characterOffset: $0.0.location))
         }
+    }
+
+    private func contentForToken(token: SyntaxToken, file: File) -> String {
+        return file.contents.substringWithByteRange(start: token.offset,
+                                                    length: token.length) ?? ""
     }
 }